re-enable history syncing and add more tests
