+++ Tests/XMTPTests/ConversationTests.swift
@@ -189,123 +189,4 @@ class ConversationTests: XCTestCase {
 
 		await fulfillment(of: [expectation1], timeout: 3)
 	}
-
-	func testSyncConsent() async throws {
-		let fixtures = try await fixtures()
-
-		let key = try Crypto.secureRandomBytes(count: 32)
-		let alix = try PrivateKey.generate()
-		var alixClient = try await Client.create(
-			account: alix,
-			options: .init(
-				api: .init(env: .local, isSecure: false),
-				dbEncryptionKey: key,
-				dbDirectory: "xmtp_db"
-			)
-		)
-
-		let dm = try await alixClient.conversations.findOrCreateDm(
-			with: fixtures.bo.walletAddress)
-		try await dm.updateConsentState(state: .denied)
-		XCTAssertEqual(try dm.consentState(), .denied)
-
-		try await fixtures.boClient.conversations.sync()
-		let boDm = try await fixtures.boClient.findConversation(conversationId: dm.id)
-
-		var alixClient2 = try await Client.create(
-			account: alix,
-			options: .init(
-				api: .init(env: .local, isSecure: false),
-				dbEncryptionKey: key,
-				dbDirectory: "xmtp_db2"
-			)
-		)
-
-		let state = try await alixClient2.inboxState(refreshFromNetwork: true)
-		XCTAssertEqual(state.installations.count, 2)
-
-		try await fixtures.boClient.conversations.sync()
-		try await boDm?.sync()
-		try await alixClient2.preferences.syncConsent()
-		try await alixClient.conversations.syncAllConversations()
-		sleep(2)
-		try await alixClient2.conversations.syncAllConversations()
-		sleep(2)
-
-		if let dm2 = try await alixClient2.findConversation(conversationId: dm.id) {
-			XCTAssertEqual(try dm2.consentState(), .denied)
-
-			try await alixClient2.preferences.setConsentState(
-				entries: [
-					ConsentRecord(
-						value: dm2.id,
-						entryType: .conversation_id,
-						consentType: .allowed
-					)
-				]
-			)
-			let convoState = try await alixClient2.preferences
-				.conversationState(
-					conversationId: dm2.id)
-			XCTAssertEqual(convoState, .allowed)
-			XCTAssertEqual(try dm2.consentState(), .allowed)
-		}
-	}
-	
-	func testStreamConsent() async throws {
-		let fixtures = try await fixtures()
-
-		let key = try Crypto.secureRandomBytes(count: 32)
-		let alix = try PrivateKey.generate()
-
-		let alixClient = try await Client.create(
-			account: alix,
-			options: .init(
-				api: .init(env: .local, isSecure: false),
-				dbEncryptionKey: key,
-				dbDirectory: "xmtp_db"
-			)
-		)
-
-		let alixGroup = try await alixClient.conversations.newGroup(with: [fixtures.bo.walletAddress])
-
-		let alixClient2 = try await Client.create(
-			account: alix,
-			options: .init(
-				api: .init(env: .local, isSecure: false),
-				dbEncryptionKey: key,
-				dbDirectory: "xmtp_db2"
-			)
-		)
-		
-		try await alixGroup.send(content: "Hello")
-		try await alixClient.conversations.syncAllConversations()
-		try await alixClient2.conversations.syncAllConversations()
-		let alixGroup2 = try alixClient2.findGroup(groupId: alixGroup.id)!
-
-		var consentList = [ConsentRecord]()
-		let expectation = XCTestExpectation(description: "Stream Consent")
-		expectation.expectedFulfillmentCount = 3
-
-		Task(priority: .userInitiated) {
-			for try await entry in await alixClient.preferences.streamConsent() {
-				consentList.append(entry)
-				expectation.fulfill()
-			}
-		}
-		sleep(1)
-		try await alixGroup2.updateConsentState(state: .denied)
-		let dm = try await alixClient2.conversations.newConversation(with: fixtures.caro.walletAddress)
-		try await dm.updateConsentState(state: .denied)
-
-		sleep(5)
-		try await alixClient.conversations.syncAllConversations()
-		try await alixClient2.conversations.syncAllConversations()
-
-		await fulfillment(of: [expectation], timeout: 3)
-		print(consentList)
-		XCTAssertEqual(try alixGroup.consentState(), .denied)
-	}
-
-
 }
